Move module declaration into definition file, override PageProps and add template macro types
